requirements.txt
beautifulsoup4==4.12.0
folium==0.12.1.post1
mysql_connector_repackaged==0.3.1
pandas==1.4.3
pdfcrowd==5.11.0
price_parser==0.3.4
Shapely==1.7.1
streamlit==1.14.0
streamlit_folium==0.6.15
